Split hunter and non hunter logic into .h and .cpp

+++ include/CellsEvo/core/app.h
@@ -24,7 +24,7 @@ class App {
       int window_width,
       int window_height,
       int fps,
-      int food_production_rate
+      float food_production_rate
   );
   ~App();
